requirements.txt
tensorflow-gpu==1.5
keras==2.0.8
numpy
scipy
Pillow
cython
matplotlib
scikit-image
opencv-python
h5py=2.6.0
imgaug
IPython[all]
tqdm